controller e service Cliente

+++ src/main/java/com/example/lcreparos/services/ClienteService.java
@@ -5,6 +5,8 @@ import java.util.stream.Collectors;
 
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import com.example.lcreparos.Dtos.ClienteDto;
@@ -18,37 +20,75 @@ public class ClienteService {
     private ClienteRepository clienteRepository;
     @Autowired
     private ModelMapper modelMapper;
-    
-    public List<ClienteDto> findAll(){
+
+    public List<ClienteDto> findAll() {
+
         List<Cliente> lista = clienteRepository.findAll();
+
         List<ClienteDto> listaDto = lista.stream().map(cliente -> modelMapper.map(cliente, ClienteDto.class))
                 .collect(Collectors.toList());
+
         return listaDto;
     }
-    public ClienteDto findById(Long id){
-        Cliente cliente = clienteRepository.findById(id).orElse(null);
-        ClienteDto clienteDto = modelMapper.map(cliente, ClienteDto.class);
 
-        
-        return clienteDto;
+    public Page<Cliente> findAllPage(Pageable pageable) {
+
+        Page<Cliente> pageAll = clienteRepository.findAll(pageable);
+
+        return pageAll;
     }
 
-    public void saveCliente(Cliente cliente){
-        clienteRepository.save(cliente);
-    }
+    public ClienteDto findById(Long id) {
 
-    public void deleteCliente(Long id){
-        clienteRepository.deleteById(id);
-    }
+        boolean existe = clienteRepository.existsById(id);
 
-    public ClienteDto updateCliente(Cliente cliente){
-        boolean existe = clienteRepository.existsById(cliente.getIdCliente());
+        if (existe) {
 
-        if (existe == true) {
-            saveCliente(cliente);
+            Cliente cliente = clienteRepository.findById(id).orElse(null);
             ClienteDto clienteDto = modelMapper.map(cliente, ClienteDto.class);
             return clienteDto;
         }
+
+        return new ClienteDto();
+    }
+
+    public Cliente saveCliente(Cliente cliente) {
+
+        try{
+
+            clienteRepository.save(cliente);
+            return cliente;
+        }
+        catch(Exception e){
+
+            return new Cliente();
+        }
+    }
+
+    public boolean deleteCliente(Long id) {
+
+        boolean existe = clienteRepository.existsById(id);
+
+        if (existe) {
+
+            clienteRepository.deleteById(id);
+            return true;
+        }
+
+        return false;
+    }
+
+    public ClienteDto updateCliente(Cliente cliente) {
+
+        boolean existe = clienteRepository.existsById(cliente.getIdCliente());
+
+        if (existe == true) {
+
+            saveCliente(cliente);
+            ClienteDto clienteDto = modelMapper.map(cliente, ClienteDto.class);
+            return clienteDto;            
+        }
+        
         return null;
     }
 }
